refactor: split main in lib + main plus logical refactoring
