+++ src/commands/package.rs
@@ -5,12 +5,10 @@ use std::{
     process::{Command, Stdio},
 };
 
+use crate::execution::{setup_ctrlc_hook, ExecutionStrategy, Executor};
+use crate::options::KerblamTomlOptions;
 use crate::options::Pipe;
 use crate::utils::find_files;
-use crate::{
-    commands::run::{setup_ctrlc_hook, ExecutionStrategy, Executor},
-    options::KerblamTomlOptions,
-};
 
 use anyhow::{bail, Result};
